Adjust time formatting in spreadsheet
Moment documentation indicates s and m are for seconds and minutes but doesn't really define a difference other than it looked like it was for how granular you wanted it.  Changing m to mm and s to ss on line 395 makes the time more readable.  Previously it might end up as 1/24/2020 3:2:5 as an example.  Now that same stamp will come through as 1/24/2020 3:02:05.